Refactor #33 : 블록체인 요청 기록 Entity 변환 클래스 추가
+ 클래스의 단일 책임 원칙을 위해 분리
